Only write protocolflags to text if protocol 999
It is only read from the demo for protocol 999, but then unconditionally
written to the text file. That seems to have been overlooked in 300970b
because this causes the generated text file to be invalid for all other
protocol versions.

Adding that conditional and bumping version number for bugfix release.